Fixed dir name for ulr wildcard
